backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
cloudinary==1.36.0
razorpay==1.4.2
Pillow==10.1.0
django-cloudinary-storage==0.3.0

